requirements.txt
django==5.0.4
psycopg2-binary>=2.9
django-allauth[socialaccount]==0.56.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth[with_social]==6.0.0
django-cors-headers==4.3.1
pre-commit==3.7.1
